# contracts/openapi/public-api.yaml
openapi: 3.1.0
info:
  title: Gereh Public API
  version: 0.1.0
  description: |
    Public REST API exposed by the Gereh API BFF.
servers:
  - url: /
...
paths:
...
  /v1/tenants/{tenantId}/tasks/{taskId}/dependencies:
    post:
      operationId: addTaskDependency
      summary: Add a task dependency
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTaskDependencyRequest"
      responses:
        "201":
          description: Dependency created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddTaskDependencyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/tenants/{tenantId}/tasks/{taskId}/dependencies/{dependsOnTaskId}:
    delete:
      operationId: removeTaskDependency
      summary: Remove a task dependency
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: dependsOnTaskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Dependency removed.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/tenants/{tenantId}/tasks/{taskId}/assignments:
    post:
      operationId: assignTask
      summary: Assign a task
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignTaskRequest"
      responses:
        "201":
          description: Assignment created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignTaskResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/tenants/{tenantId}/tasks/{taskId}/assignments/{assignmentId}:
    delete:
      operationId: unassignTask
      summary: Unassign a task
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: assignmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Assignment removed.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/tenants/{tenantId}/tasks/{taskId}/comments:
    post:
      operationId: addComment
      summary: Add a comment
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCommentRequest"
      responses:
        "201":
          description: Comment created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddCommentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/tenants/{tenantId}/tasks/{taskId}/comments/{commentId}:
    patch:
      operationId: updateComment
      summary: Update a comment
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: commentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommentRequest"
      responses:
        "200":
          description: Updated comment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateCommentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteComment
      summary: Delete a comment
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: commentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteCommentRequest"
      responses:
        "200":
          description: Comment deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/tenants/{tenantId}/tasks/{taskId}/artifacts:
    post:
      operationId: addArtifact
      summary: Add an artifact
      description: Records artifact metadata for a task.
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddArtifactRequest"
      responses:
        "201":
          description: Artifact recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddArtifactResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/tenants/{tenantId}/tasks/{taskId}/artifacts/{artifactId}:
    delete:
      operationId: deleteArtifact
      summary: Delete an artifact
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: artifactId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Artifact deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/tenants/{tenantId}/tasks/{taskId}/checklist:
    post:
      operationId: addChecklistItem
      summary: Add a checklist item
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddChecklistItemRequest"
      responses:
        "201":
          description: Checklist item created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddChecklistItemResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/tenants/{tenantId}/tasks/{taskId}/checklist/{itemId}:
    patch:
      operationId: updateChecklistItem
      summary: Update a checklist item
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChecklistItemRequest"
      responses:
        "200":
          description: Updated checklist item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateChecklistItemResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteChecklistItem
      summary: Delete a checklist item
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteChecklistItemRequest"
      responses:
        "200":
          description: Checklist item deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/tenants/{tenantId}/tasks/{taskId}/schedule:
    put:
      operationId: upsertTaskSchedule
      summary: Create or update a task schedule
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertTaskScheduleRequest"
      responses:
        "200":
          description: Schedule stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpsertTaskScheduleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteTaskSchedule
      summary: Delete a task schedule
      tags:
        - Work
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Schedule deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
components:
...
  schemas:
...
    AddTaskDependencyRequest:
      type: object
      additionalProperties: false
      required:
        - dependsOnTaskId
      properties:
        dependsOnTaskId:
          type: string
          format: uuid
    AddTaskDependencyResponse:
      type: object
      additionalProperties: false
      required:
        - dependency
      properties:
        dependency:
          $ref: "#/components/schemas/TaskDependency"
...
    AssignTaskRequest:
      type: object
      additionalProperties: false
      required:
        - assigneeType
        - role
      properties:
        assigneeType:
          type: string
          enum:
            - ASSIGNEE_TYPE_USER
            - ASSIGNEE_TYPE_AGENT
        userId:
          type:
            - string
            - "null"
          format: uuid
        agentId:
          type:
            - string
            - "null"
          format: uuid
        role:
          type: string
          enum:
            - ASSIGNMENT_ROLE_PRIMARY
            - ASSIGNMENT_ROLE_COLLABORATOR
            - ASSIGNMENT_ROLE_REVIEWER
    AssignTaskResponse:
      type: object
      additionalProperties: false
      required:
        - assignment
      properties:
        assignment:
          $ref: "#/components/schemas/TaskAssignment"
...
    AddCommentRequest:
      type: object
      additionalProperties: false
      required:
        - authorType
        - body
      properties:
        authorType:
          type: string
          enum:
            - COMMENT_AUTHOR_TYPE_USER
            - COMMENT_AUTHOR_TYPE_AGENT
        authorUserId:
          type:
            - string
            - "null"
          format: uuid
        authorAgentId:
          type:
            - string
            - "null"
          format: uuid
        body:
          type: string
          minLength: 1
          maxLength: 4000
    AddCommentResponse:
      type: object
      additionalProperties: false
      required:
        - comment
      properties:
        comment:
          $ref: "#/components/schemas/Comment"
    UpdateCommentRequest:
      type: object
      additionalProperties: false
      required:
        - expectedVersion
        - body
      properties:
        expectedVersion:
          type: integer
          format: int64
        body:
          type: string
          minLength: 1
          maxLength: 4000
...
  responses:
    BadRequest:
      description: The request was malformed or missing required fields.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication is required.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The actor lacks the required permission.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The requested resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The request conflicts with the current resource state.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
...